pyproject.toml
[tool.poetry]
name = "pycalculator"
version = "0.1.0"
description = ""
authors = ["Wittelsbach-Konig <kiryuschin.vitalij@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
antlr4-python3-runtime = "^4.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
